rename lambdas and make the clear lambda folder strucutre
