+++ README.md
@@ -10,10 +10,13 @@ AWS CDK project for email validation and user registration with SSO (Google, Fac
 
 1. npm install
 2. npm run build
-3. npx cdk deploy --profile {your_aws_profile_with_good_permissions} --parameters DomainPrefix={uniq_domain_prefix_for_cognito}
+3. npx cdk deploy --profile {your_aws_profile_with_good_permissions} --parameters
+   DomainPrefix={uniq_domain_prefix_for_cognito}
 
 ## Usage
+
 - Request to start validation
+
 ```shell
 POST https://{your_gw_endpoint}.execute-api.us-east-1.amazonaws.com/prod/validate
 Authorization: {id_token_from_ouauth(hosted_UI cognito section)}
@@ -23,19 +26,25 @@ Content-Type: application/json
 "email": "vladbbk@hotmail.com"
 }
 ```
+
 example of response
+
 ```shell
 {
   "requestId": "c4a8c3d9-4e10-433f-8bbd-de1aa61de018"
 }
 ```
+
 - Check the status of validation and result
+
 ```shell
 GET https://k40vx6ypq7.execute-api.us-east-1.amazonaws.com/prod/validate/c4a8c3d9-4e10-433f-8bbd-de1aa61de018
 Authorization: {id_token_from_ouauth(hosted_UI cognito section)}
 Accept: application/json
 ```
+
 example of response
+
 ```shell
 {
   "valid": false,
@@ -46,6 +55,7 @@ example of response
 ```
 
 Used tools:
+
 - AWS CDK
 - AWS Cognito
 - AWS SQS
@@ -56,19 +66,22 @@ Used tools:
 
 Structure:
 This is Default CDK template but with some changes.
+
 ```shell 
 |-/lambda - folder contains separate directories for each Lambda function and shared services,interfaces
-  |--/check-status - a lambda function that called by API GW and return the status of validation and result
-  |--/initiate-validation - a lambda that called by API GW and return requestId also initiate other steps like send to SQS and save to db
-  |--/sqs-to-stepfunction - a lambda that triggered by SQS and start the AWS Step function execution
-  |--/user-registration - a custom lambda to save user to db (Risht now not used)
-  |--/validators/ - a folder with lamdas that responsive for validation email and used in Step Function
-     |--/mx - a lambda that validate MX record
-     |--/cname - a lambda that validate CNAME
-     |--/aggreate-result - a lambda that aggreagte result from lambdas above and save to db final result
-  |--/interfaces - a shared folder with interdaces for all lamdas
-  |--/services - a shared folder with services for all lamdas
-     |--/dynamo.service
-     |--/email.service
+  |--/functions/ - folder contains separate directories for each Lambda function
+      |---/validation/ - this folder groups all validation-related Lambda functions together
+          |----/check-status - a lambda function that called by API GW and return the status of validation and result
+          |----/initiate-validation - a lambda that called by API GW and return requestId also initiate other steps like send to SQS and save to db
+          |----/sqs-to-stepfunction - a lambda that triggered by SQS and start the AWS Step function execution
+          |----/email-steps/ - a folder with lamdas that responsive for validation email and used in Step Function
+              |-----/mx-validator - a lambda that validate MX record
+              |-----/cname-validator - a lambda that validate CNAME
+              |-----/result-aggregator - a lambda that aggreagte result from lambdas above and save to db final result
+  |--/shared/ - folder contains reusable components like interfaces, services, and utilities, promoting DRY 
+      |---/interfaces - a shared folder with interdaces for all lamdas
+      |---/services - a shared folder with services for all lamdas
+         |----/dynamo.service
+         |----/email.service
 ```
