+++ lambda/functions/validation/sqs-to-stepfunction/handler.ts
