+++ lambda/shared/exceptions/email-not-found.exception.ts
