+++ lambda/shared/interfaces/email-validator.interfaces.ts
