+++ lambda/shared/interfaces/index.ts
